+++ braintree/webhook_testing_gateway.py
@@ -62,6 +62,10 @@ class WebhookTestingGateway(object):
             return self.__subscription_charged_successfully_sample_xml(id)
         elif kind == WebhookNotification.Kind.AccountUpdaterDailyReport:
             return self.__account_updater_daily_report_sample_xml()
+        elif kind == WebhookNotification.Kind.IdealPaymentComplete:
+            return self.__ideal_payment_complete_sample_xml(id)
+        elif kind == WebhookNotification.Kind.IdealPaymentFailed:
+            return self.__ideal_payment_failed_sample_xml(id)
         else:
             return self.__subscription_sample_xml(id)
 
@@ -173,6 +177,24 @@ class WebhookTestingGateway(object):
         """ % id
 
     def __dispute_opened_sample_xml(self, id):
+        if id == "legacy_dispute_id":
+            return self.__old_dispute_opened_sample_xml(id)
+        else:
+            return self.__new_dispute_opened_sample_xml(id)
+
+    def __dispute_lost_sample_xml(self, id):
+        if id == "legacy_dispute_id":
+            return self.__old_dispute_lost_sample_xml(id)
+        else:
+            return self.__new_dispute_lost_sample_xml(id)
+
+    def __dispute_won_sample_xml(self, id):
+        if id == "legacy_dispute_id":
+            return self.__old_dispute_won_sample_xml(id)
+        else:
+            return self.__new_dispute_won_sample_xml(id)
+
+    def __old_dispute_opened_sample_xml(self, id):
         return """
             <dispute>
               <amount>250.00</amount>
@@ -191,7 +213,7 @@ class WebhookTestingGateway(object):
             </dispute>
         """ % (id, id)
 
-    def __dispute_lost_sample_xml(self, id):
+    def __old_dispute_lost_sample_xml(self, id):
         return """
             <dispute>
               <amount>250.00</amount>
@@ -210,7 +232,7 @@ class WebhookTestingGateway(object):
             </dispute>
         """ % (id, id)
 
-    def __dispute_won_sample_xml(self, id):
+    def __old_dispute_won_sample_xml(self, id):
         return """
             <dispute>
               <amount>250.00</amount>
@@ -228,6 +250,168 @@ class WebhookTestingGateway(object):
               <date-opened type="date">2014-03-28</date-opened>
               <date-won type="date">2014-09-01</date-won>
             </dispute>
+        """ % (id, id)
+
+    def __new_dispute_opened_sample_xml(self, id):
+        return """
+        <dispute>
+          <id>%s</id>
+          <amount>100.00</amount>
+          <amount-disputed>100.00</amount-disputed>
+          <amount-won>95.00</amount-won>
+          <case-number>CASE-12345</case-number>
+          <created-at type="datetime">2017-06-16T20:44:41Z</created-at>
+          <currency-iso-code>USD</currency-iso-code>
+          <forwarded-comments nil="true"/>
+          <kind>chargeback</kind>
+          <merchant-account-id>ytnlulaloidoqwvzxjrdqputg</merchant-account-id>
+          <reason>fraud</reason>
+          <reason-code nil="true"/>
+          <reason-description nil="true"/>
+          <received-date type="date">2016-02-15</received-date>
+          <reference-number>REF-9876</reference-number>
+          <reply-by-date type="date">2016-02-22</reply-by-date>
+          <status>open</status>
+          <updated-at type="datetime">2017-06-16T20:44:41Z</updated-at>
+          <original-dispute-id>9qde5qgp</original-dispute-id>
+          <status-history type="array">
+            <status-history>
+              <status>open</status>
+              <timestamp type="datetime">2017-06-16T20:44:41Z</timestamp>
+            </status-history>
+          </status-history>
+          <evidence type="array"/>
+          <transaction>
+            <id>%s</id>
+            <amount>100.00</amount>
+            <created-at>2017-06-21T20:44:41Z</created-at>
+            <order-id nil="true"/>
+            <purchase-order-number nil="true"/>
+            <payment-instrument-subtype>Visa</payment-instrument-subtype>
+          </transaction>
+          <date-opened type=\"date\">2014-03-28</date-opened>
+        </dispute>
+        """ % (id, id)
+
+    def __new_dispute_lost_sample_xml(self, id):
+        return """
+        <dispute>
+          <id>%s</id>
+          <amount>100.00</amount>
+          <amount-disputed>100.00</amount-disputed>
+          <amount-won>95.00</amount-won>
+          <case-number>CASE-12345</case-number>
+          <created-at type="datetime">2017-06-16T20:44:41Z</created-at>
+          <currency-iso-code>USD</currency-iso-code>
+          <forwarded-comments nil="true"/>
+          <kind>chargeback</kind>
+          <merchant-account-id>ytnlulaloidoqwvzxjrdqputg</merchant-account-id>
+          <reason>fraud</reason>
+          <reason-code nil="true"/>
+          <reason-description nil="true"/>
+          <received-date type="date">2016-02-15</received-date>
+          <reference-number>REF-9876</reference-number>
+          <reply-by-date type="date">2016-02-22</reply-by-date>
+          <status>lost</status>
+          <updated-at type="datetime">2017-06-21T20:44:41Z</updated-at>
+          <original-dispute-id>9qde5qgp</original-dispute-id>
+          <status-history type="array">
+            <status-history>
+              <status>open</status>
+              <timestamp type="datetime">2017-06-16T20:44:41Z</timestamp>
+            </status-history>
+            <status-history>
+              <status>lost</status>
+              <timestamp type="datetime">2017-06-25T20:50:55Z</timestamp>
+            </status-history>
+          </status-history>
+          <evidence type="array">
+            <evidence>
+              <id>rxtngk9j5j93tsrq</id>
+              <comments nil="true"/>
+              <created-at type="datetime">2017-06-21T20:44:42Z</created-at>
+              <sent-to-processor-at nil="true"/>
+              <url>s3.amazonaws.com/foo.jpg</url>
+            </evidence>
+            <evidence>
+              <id>88cfb8dd</id>
+              <comments>text evidence</comments>
+              <created-at type="datetime">2017-06-21T20:44:42Z</created-at>
+              <sent-to-processor-at nil="true"/>
+              <url nil="true"/>
+            </evidence>
+          </evidence>
+          <transaction>
+            <id>%s</id>
+            <amount>100.00</amount>
+            <created-at>2017-06-21T20:44:41Z</created-at>
+            <order-id nil="true"/>
+            <purchase-order-number nil="true"/>
+            <payment-instrument-subtype>Visa</payment-instrument-subtype>
+          </transaction>
+          <date-opened type=\"date\">2014-03-28</date-opened>
+        </dispute>
+        """ % (id, id)
+
+    def __new_dispute_won_sample_xml(self, id):
+        return """
+        <dispute>
+          <id>%s</id>
+          <amount>100.00</amount>
+          <amount-disputed>100.00</amount-disputed>
+          <amount-won>95.00</amount-won>
+          <case-number>CASE-12345</case-number>
+          <created-at type="datetime">2017-06-16T20:44:41Z</created-at>
+          <currency-iso-code>USD</currency-iso-code>
+          <forwarded-comments nil="true"/>
+          <kind>chargeback</kind>
+          <merchant-account-id>ytnlulaloidoqwvzxjrdqputg</merchant-account-id>
+          <reason>fraud</reason>
+          <reason-code nil="true"/>
+          <reason-description nil="true"/>
+          <received-date type="date">2016-02-15</received-date>
+          <reference-number>REF-9876</reference-number>
+          <reply-by-date type="date">2016-02-22</reply-by-date>
+          <status>won</status>
+          <updated-at type="datetime">2017-06-21T20:44:41Z</updated-at>
+          <original-dispute-id>9qde5qgp</original-dispute-id>
+          <status-history type="array">
+            <status-history>
+              <status>open</status>
+              <timestamp type="datetime">2017-06-16T20:44:41Z</timestamp>
+            </status-history>
+            <status-history>
+              <status>won</status>
+              <timestamp type="datetime">2017-06-25T20:50:55Z</timestamp>
+            </status-history>
+          </status-history>
+          <evidence type="array">
+            <evidence>
+              <id>rxtngk9j5j93tsrq</id>
+              <comments nil="true"/>
+              <created-at type="datetime">2017-06-21T20:44:42Z</created-at>
+              <sent-to-processor-at nil="true"/>
+              <url>s3.amazonaws.com/foo.jpg</url>
+            </evidence>
+            <evidence>
+              <id>88cfb8dd</id>
+              <comments>text evidence</comments>
+              <created-at type="datetime">2017-06-21T20:44:42Z</created-at>
+              <sent-to-processor-at nil="true"/>
+              <url nil="true"/>
+            </evidence>
+          </evidence>
+          <transaction>
+            <id>%s</id>
+            <amount>100.00</amount>
+            <created-at>2017-06-21T20:44:41Z</created-at>
+            <order-id nil="true"/>
+            <purchase-order-number nil="true"/>
+            <payment-instrument-subtype>Visa</payment-instrument-subtype>
+          </transaction>
+          <date-opened type=\"date\">2014-03-28</date-opened>
+          <date-won type=\"date\">2014-09-01</date-won>
+        </dispute>
         """ % (id, id)
 
     def __subscription_sample_xml(self, id):
@@ -345,3 +529,33 @@ class WebhookTestingGateway(object):
                 <report-url>link-to-csv-report</report-url>
             </account-updater-daily-report>
             """
+
+    def __ideal_payment_complete_sample_xml(self, id):
+        return """
+            <ideal-payment>
+                <id>%s</id>
+                <status>COMPLETE</status>
+                <issuer>ABCISSUER</issuer>
+                <order-id>ORDERABC</order-id>
+                <currency>EUR</currency>
+                <amount>10.00</amount>
+                <created-at>2016-11-29T23:27:34.547Z</created-at>
+                <approval-url>https://example.com</approval-url>
+                <ideal-transaction-id>1234567890</ideal-transaction-id>
+            </ideal-payment>
+            """ % id
+
+    def __ideal_payment_failed_sample_xml(self, id):
+        return """
+            <ideal-payment>
+                <id>%s</id>
+                <status>FAILED</status>
+                <issuer>ABCISSUER</issuer>
+                <order-id>ORDERABC</order-id>
+                <currency>EUR</currency>
+                <amount>10.00</amount>
+                <created-at>2016-11-29T23:27:34.547Z</created-at>
+                <approval-url>https://example.com</approval-url>
+                <ideal-transaction-id>1234567890</ideal-transaction-id>
+            </ideal-payment>
+            """ % id